+++ app/layout.tsx
@@ -30,8 +30,18 @@ export default function RootLayout({children,}: Readonly<{
 
   return (
     <html lang="en">
+    <head>
+      <link
+        rel="preload"
+        href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined"
+        as="style"
+      />
+      <link
+        href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined"
+        rel="stylesheet"
+      />
+    </head>
     <body className={[poppins_init.className, inter_init.className].join(' ')}>
-
     <a href="#mainContent" className="skip-nav">Skip to main content</a>
     <AuthProvider>
       <Header/>