+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Added existing header and footer from production to app [#89]
 - Added two new TypeAhead/Autosuggest components, one with and one without the "Other" option [#10]
 - Added the use of graphql-codegen to generate types for graphql requests [#10]
+- Added Material Symbols and a custom `DmpIcon` component that will render a specific icon using it's short-code. [#80]
+- Added the initial version of our custom ReMirror text editor [#74].
 
 ### Updates
 - Updated endpoints for 'signin' and 'signup' to be 'apollo-signin' and 'apollo-signup' [#99]